remove dependence of tree on header
